+++ station/lib/model/fuel_type.dart
@@ -1,0 +1,16 @@
+
+enum FuelType {
+  unknown,
+  petrol,
+  e5,
+  e10,
+  petrolSuperPlus,
+  petrolShellPower,
+  petrolAralUltimate,
+  diesel,
+  dieselHvo100,
+  dieselTruck,
+  dieselShellPower,
+  dieselAralUltimate,
+  lpg,
+}