+++ station/lib/model/marker_model.dart
@@ -1,6 +1,6 @@
 import 'package:navigation/coordinate_model.dart';
 import 'package:station/model/currency_model.dart';
-import 'package:station/model/price_model.dart';
+import 'package:station/model/fuel_type.dart';
 
 class MarkerModel {
   final int id;